+++ common/models/I18n.js
@@ -33,24 +33,27 @@ const logger	= reqlocal(path.join('server', 'boot', 'winston.js')).logger;
  */
 module.exports = function(I18n) {
 
-	// https://loopback.io/doc/en/lb3/Authentication-authorization-and-permissions.html
-	I18n.disableRemoteMethodByName('upsert');								// disables PATCH /I18ns
-	I18n.disableRemoteMethodByName('find');									// disables GET /I18ns
-	I18n.disableRemoteMethodByName('replaceOrCreate');						// disables PUT /I18ns
-	I18n.disableRemoteMethodByName('create');								// disables POST /I18ns
-	I18n.disableRemoteMethodByName('prototype.updateAttributes');			// disables PATCH /I18ns/{id}
-	I18n.disableRemoteMethodByName('findById');								// disables GET /I18ns/{id}
-	I18n.disableRemoteMethodByName('exists');								// disables HEAD /I18ns/{id}
-	I18n.disableRemoteMethodByName('replaceById');							// disables PUT /I18ns/{id}
-	I18n.disableRemoteMethodByName('deleteById');							// disables DELETE /I18ns/{id}
-	I18n.disableRemoteMethodByName('prototype.__findById__accessTokens');	// disable GET /I18ns/{id}/accessTokens/{fk}
-	I18n.disableRemoteMethodByName('prototype.__updateById__accessTokens');	// disable PUT /I18ns/{id}/accessTokens/{fk}
-	I18n.disableRemoteMethodByName('prototype.__destroyById__accessTokens');// disable DELETE /I18ns/{id}/accessTokens/{fk}
-	I18n.disableRemoteMethodByName('prototype.__count__accessTokens');		// disable  GET /I18ns/{id}/accessTokens/count
-	I18n.disableRemoteMethodByName('count');								// disables GET /I18ns/count
-	I18n.disableRemoteMethodByName('findOne');								// disables GET /I18ns/findOne
-	I18n.disableRemoteMethodByName('update');								// disables POST /I18ns/update
-	I18n.disableRemoteMethodByName('upsertWithWhere');						// disables POST /I18ns/upsertWithWhere
+	if (process.env.NODE_ENV !== undefined) {
+		// https://loopback.io/doc/en/lb3/Authentication-authorization-and-permissions.html
+		I18n.disableRemoteMethodByName('upsert');								// disables PATCH /I18ns
+		I18n.disableRemoteMethodByName('find');									// disables GET /I18ns
+		I18n.disableRemoteMethodByName('replaceOrCreate');						// disables PUT /I18ns
+		I18n.disableRemoteMethodByName('create');								// disables POST /I18ns
+		I18n.disableRemoteMethodByName('prototype.updateAttributes');			// disables PATCH /I18ns/{id}
+		I18n.disableRemoteMethodByName('findById');								// disables GET /I18ns/{id}
+		I18n.disableRemoteMethodByName('exists');								// disables HEAD /I18ns/{id}
+		I18n.disableRemoteMethodByName('replaceById');							// disables PUT /I18ns/{id}
+		I18n.disableRemoteMethodByName('deleteById');							// disables DELETE /I18ns/{id}
+		I18n.disableRemoteMethodByName('prototype.__findById__accessTokens');	// disable GET /I18ns/{id}/accessTokens/{fk}
+		I18n.disableRemoteMethodByName('prototype.__updateById__accessTokens');	// disable PUT /I18ns/{id}/accessTokens/{fk}
+		I18n.disableRemoteMethodByName('prototype.__destroyById__accessTokens');// disable DELETE /I18ns/{id}/accessTokens/{fk}
+		I18n.disableRemoteMethodByName('prototype.__count__accessTokens');		// disable  GET /I18ns/{id}/accessTokens/count
+		I18n.disableRemoteMethodByName('createChangeStream');					// disables POST /I18ns/change-stream
+		I18n.disableRemoteMethodByName('count');								// disables GET /I18ns/count
+		I18n.disableRemoteMethodByName('findOne');								// disables GET /I18ns/findOne
+		I18n.disableRemoteMethodByName('update');								// disables POST /I18ns/update
+		I18n.disableRemoteMethodByName('upsertWithWhere');						// disables POST /I18ns/upsertWithWhere
+	}
 
 	/**
 	 * Get the list of supported languages for GUI translations
